Refactored and fixed bugs in herbstluftwm scripts
Issue of herbst crashing sometimes has been attributed to concurrency issue with
lock and unlock commands. With most scripts being launch asyncrounsly this is
very hard to account for. The fix was to remove as many calls to lock and
unlock as possible, and to do this while still retaing the same visual
coherence. Some calss to lock and unlock remain, but they should be fairly
innocuous.

Desktop windows now prioritize the a focussed empty frame over an inorder list of empty
frames when moving windows.

Most window moving functionality has been abstracted to hc-lib.sh

find-focus.pl adds a window to an empty focussed frame in a given layout. Useful
for avoiding calls to lock and unlock.