Changed how the lemonbar is notified of a new window
The lemonbar is now only notified of a new window when the new window is
moved to the hidden tag (as opposed to an existing empty frame). When
notifying the lemonbar of a new window we request the notification animation
to alert the user that there is a new window.